+++ wp-plugin-boilerplate.php
@@ -3,7 +3,7 @@
  * Plugin Name: WP Plugin Boilerplate
  * Description: An opinionated, OOP-first WordPress plugin boilerplate for building long-lived, maintainable plugins.
  * Plugin URI: https://www.ulhas.net/labs/wp-plugin-boilerplate/
- * Version: 1.1.1
+ * Version: 1.1.2
  * Author: Ulhas Vardhan Golchha
  * Author URI: https://www.ulhas.net/
  * Text Domain: wp-plugin-boilerplate
@@ -12,7 +12,6 @@
 defined('ABSPATH') || exit;
 
 if ( ! file_exists( __DIR__ . '/vendor/autoload.php' ) ) return; // Bundled dependencies are required
-
 
 /**
  * -------------------------------------------------------------------------
@@ -27,7 +26,7 @@ if (!defined('WPPB_PREFIX')) {
 }
 
 if (!defined('WPPB_VERSION')) {
-	define('WPPB_VERSION', '1.1.1');
+	define('WPPB_VERSION', '1.1.2');
 }
 
 if (!defined('WPPB_SLUG')) {